refactor: added RoomForm for the room creation and join room
